requirements.txt
aiohttp==3.7.4.post0
aioice==0.7.5
aiortc==1.2.0
async-timeout==3.0.1
attrs==20.3.0
av==8.0.3
bidict==0.21.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
crc32c==2.2
cryptography==3.4.7
dataclasses==0.8
dnspython==2.1.0
idna==2.10
idna-ssl==1.1.0
multidict==5.1.0
netifaces==0.10.9
numpy==1.19.5
opencv-python==4.5.1.48
pycparser==2.20
pyee==8.1.0
pylibsrtp==0.6.8
python-engineio==4.1.0
python-socketio==5.2.1
requests==2.25.1
typing-extensions==3.7.4.3
urllib3==1.26.4
yarl==1.6.3
